fix(internal/text): make FormatSize take int64 and clamp at largest unit
Avoids an out-of-range panic for byte counts at or above one exabyte, which
fit within int64, and lets the function accept the full int64 range.

+++ pkg/cmd/repo/read-dir/read_dir.go
@@ -188,7 +188,7 @@ func writeTable(io *iostreams.IOStreams, dir *repoDir) error {
 
 		size := "-"
 		if e.Type == "file" || e.Type == "symlink" {
-			size = text.FormatSize(e.Size)
+			size = text.FormatSize(int64(e.Size))
 		}
 
 		tp.AddField(entryType)
